requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
Pillow
python-multipart
